requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.16.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.57.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mod_wsgi==5.0.2
mysqlclient==2.2.7
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipenv==2024.4.1
platformdirs==4.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-fuzzy==0.5.0
scikit-learn==1.7.0
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
virtualenv==20.29.3
Werkzeug==3.1.3
